Task 2: only the owner of the image can edit/delete the image implemented

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -8,11 +8,9 @@ import ImageHoster.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
@@ -46,11 +44,14 @@ public class ImageController {
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
     @RequestMapping("/images/{id}/{title}")
-    public String showImage(@PathVariable("id") Integer id, @PathVariable("title") String title, Model model) {
+    public String showImage(@PathVariable("id") Integer id, @PathVariable("title") String title, Model model, HttpSession session) {
         //Image image = imageService.getImageByTitle(title);
         Image image = imageService.getImage(id);
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
+        //Debug lines added to check if added attributes are visible
+//        System.out.println("editError: "+ model.containsAttribute("editError"));
+//        System.out.println("deleteError: " +model.containsAttribute("deleteError"));
         return "images/image";
     }
 
@@ -93,9 +94,18 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session, RedirectAttributes redirectAttributes) {
         Image image = imageService.getImage(imageId);
-
+        User loggedUser = (User) session.getAttribute("loggeduser");
+        boolean isEntitled=false;
+        if(image.getUser().getId()==loggedUser.getId()){
+            isEntitled=true;
+        }
+        if(!isEntitled){
+            String error = "Only the owner of the image can edit the image";
+            redirectAttributes.addFlashAttribute("editError", error);
+            return "redirect:/images/"+image.getId()+"/"+image.getTitle();
+        }
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
@@ -133,7 +143,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" +updatedImage.getId()+"/"+ updatedImage.getTitle();
     }
 
 
@@ -141,7 +151,18 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model, RedirectAttributes redirectAttributes) {
+        Image image = imageService.getImage(imageId);
+        User loggedUser = (User) session.getAttribute("loggeduser");
+        boolean isEntitled=false;
+        if(image.getUser().getId()==loggedUser.getId()){
+            isEntitled=true;
+        }
+        if(!isEntitled){
+            String error = "Only the owner of the image can delete the image";
+            redirectAttributes.addFlashAttribute("deleteError", error);
+            return "redirect:/images/"+image.getId()+"/"+image.getTitle();
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
